ath10k:  Support board-name in fwcfg, increase BMI timer.
Some firmware needs special board files, so allow configuring
the board files per NIC in fwcfg.

And, increase BMI timer since my 9888 NIC was right on the border
of failing at about 2 seconds, and would sometimes fail to do OTP
and then of course fail to load.  It is increased to 3 seconds now.

Signed-off-by: Ben Greear <greearb@candelatech.com>