ath10k-ct:  Support better RSSI measurements.
When used with recent firmware, these changes allow the driver to
query per-chain noise-floor from the radio to better calculate the
per-chain RSSI.  The per-chain RSSI is then summed to provide the
'combined RSSI'.  This gives better per-chain RSSI as well as combined
RSSI, especially when running with more than 20Mhz bandwidths.