ath10k-ct:  Fix RSSI reporting for wave-1 firmware.
The descriptor signal values for secondary frequencies were not being
reported as expected for wave-1 hardware.  So, just look at the pri-20
value like upstream HW does.

For 4.20 and 4.19, backport more substantial fixes from the 5.4
and above kernels so that all of these report RSSI similarly.

Did not backport this to anything older than 4.19

Signed-off-by: Ben Greear <greearb@candelatech.com>